move statusCode coercion test to its own file
node-http.test.ts has a pre-existing failure (issue#4295 proxy test,
ECONNREFUSED in the gate environment) unrelated to this change.
