+++ test/js/node/http/node-http.test.ts
@@ -3587,74 +3587,6 @@ it("statusCode = 204 with a streamed write carries no body framing", async () =>
   }
 });
 
-it("statusCode assigned as a numeric string is coerced like writeHead()", async () => {
-  // Node's implicit header path routes through writeHead(this.statusCode),
-  // which coerces via |= 0 and range-checks; '404' reaches the wire as 404
-  // with the body intact, '204' still drops the body, and a non-numeric
-  // string throws ERR_HTTP_INVALID_STATUS_CODE from end().
-  const { promise: typesP, resolve: typesR } = Promise.withResolvers<[string, string]>();
-  const { promise: errP, resolve: errR } = Promise.withResolvers<unknown>();
-  const server = createServer((req, res) => {
-    if (req.url === "/404") {
-      res.statusCode = "404" as never;
-      const before = typeof (res.statusCode as unknown);
-      res.end("not found");
-      typesR([before, typeof (res.statusCode as unknown)]);
-    } else if (req.url === "/204") {
-      res.statusCode = "204" as never;
-      res.end("dropped");
-    } else if (req.url === "/204w") {
-      res.statusCode = "204" as never;
-      res.write("dropped");
-      res.end();
-    } else {
-      res.statusCode = "abc" as never;
-      try {
-        res.end("x");
-        errR(undefined);
-      } catch (e) {
-        errR(e);
-      }
-      res.destroy();
-    }
-  });
-  try {
-    server.listen(0, "127.0.0.1");
-    await once(server, "listening");
-    const { port } = server.address() as AddressInfo;
-
-    const wire = (path: string) =>
-      new Promise<string>((resolve, reject) => {
-        const socket = connect(port, "127.0.0.1");
-        let data = "";
-        socket.on("data", chunk => (data += chunk));
-        socket.on("close", () => resolve(data));
-        socket.on("error", reject);
-        socket.write(`GET ${path} HTTP/1.1\r\nHost: x\r\nConnection: close\r\n\r\n`);
-      });
-
-    const out404 = await wire("/404");
-    expect(out404).toStartWith("HTTP/1.1 404 Not Found\r\n");
-    expect(out404).toEndWith("not found");
-    expect(await typesP).toEqual(["string", "number"]);
-
-    for (const p of ["/204", "/204w"]) {
-      const out204 = await wire(p);
-      expect(out204).toStartWith("HTTP/1.1 204 No Content\r\n");
-      expect(out204).not.toContain("dropped");
-      expect(out204).not.toContain("Transfer-Encoding");
-      expect(out204).toEndWith("\r\n\r\n");
-    }
-
-    const outBad = await wire("/bad");
-    const err = (await errP) as NodeJS.ErrnoException;
-    expect(err?.code).toBe("ERR_HTTP_INVALID_STATUS_CODE");
-    expect(outBad).not.toStartWith("HTTP/1.1 200 ");
-  } finally {
-    server.close();
-  }
-});
-
 it("no 'drain' is emitted for accounting flushed after the response finished", async () => {
   let drains = 0;
   const { promise: handled, resolve: onHandled } = Promise.withResolvers<boolean>();
